Remove included_env_config and address selected style feedback

+++ lib/spack/spack/test/config.py
@@ -2300,4 +2300,5 @@ def test_included_path_git_unwritable_dest(tmp_path: pathlib.Path):
         dest.chmod(current_mode)
         # Ensure the unwritable temp dir is removed to avoid failure on some
         # platforms (e.g., rhel8).
-        shutil.rmtree(include.destination)
+        if include.destination:
+            shutil.rmtree(include.destination)  # type: ignore[arg-type]
